Update packages; fix Npgsql issues

+++ RegexBot/EntityCache/CacheChannel.cs
@@ -65,8 +65,8 @@ namespace Noikoio.RegexBot.EntityCache
                 {
                     c.CommandText = $"SELECT {QueryColumns} from {SqlHelper.TableTextChannel} WHERE "
                         + "channel_id = @Cid AND guild_id = @Gid";
-                    c.Parameters.Add("@Cid", NpgsqlTypes.NpgsqlDbType.Bigint).Value = channel;
-                    c.Parameters.Add("@Gid", NpgsqlTypes.NpgsqlDbType.Bigint).Value = guild;
+                    c.Parameters.Add("@Cid", NpgsqlTypes.NpgsqlDbType.Bigint).Value = (long)channel;
+                    c.Parameters.Add("@Gid", NpgsqlTypes.NpgsqlDbType.Bigint).Value = (long)guild;
                     c.Prepare();
                     using (var r = await c.ExecuteReaderAsync())
                     {
